Fix some strings (this commit desc is swapped with the previous)

+++ app/src/main/java/com/zacharee1/systemuituner/prefs/secure/base/BaseSecurePreference.kt
@@ -31,6 +31,6 @@ open class BaseSecurePreference(context: Context, attrs: AttributeSet) : DialogP
 
     override fun onValueChanged(newValue: Any?, key: String) {
         context.prefManager.putString(writeKey!!, newValue.toString())
-        context.writeSetting(type, writeKey, newValue.toString().toInt())
+        context.writeSetting(type, writeKey, newValue.toString())
     }
 }
